FAT: Fix warning
fs/fat/dir.c:43: warning: 'de' may be used uninitialized in this function

The complexity of the code flow makes this seem possible.  Initialize
the value to NULL to eliminate the compiler warning.  This only masks
the warning, since if the value were unused, there would still be a
NULL pointer.